Add kramdown-parser-gfm dependency
After initially installing the gems for this project, I found that the
"kramdown-parser-gfm" library was missing (`jekyll serve` failed with an
error that this library was missing).  Adding it to the `Gemfile` and
rerunning `bundle install` then allowed `jekyll serve` to work as
expected and the development version of the website could be viewed in a
web browser.