Fix link to Cookie Consent CDN
It turns out that there was an extra `>` sign in the HTML which added
the link to the Cookie Consent upstream Javascript code.  This caused
the upstream code not to be loaded (due to an HTML syntax error) which
removed an important prerequisite to getting Cookie Consent to work.
This patch fixes the `<script>` tag HTML and moves the `<script>` tag to
the after the `<footer>` tag as this is a better location for Javascript
code (i.e. right at the end of the HTML.