Add Cookie Consent to website
The missing piece of the puzzle from commit
bd07e2994dc57488450fee036ed3d4919fb75ec4 was that this Jekyll theme
works differently to the standard minimal mistakes Jekyll theme.  In
particular, there doesn't seem to be any need for a `after_scripts`
option in the configuration: one can simply extend the `footer.html`
template to point to the `cookie-consent.js` file.  This was the problem
with the commit referenced above: although the `after_scripts` option
was configured correctly, this theme was ignoring the option and hence
the `cookie-consent.js` code wasn't being included in the site and hence
the banner wasn't displaying.  I noticed a pattern for file inclusion in
the `scripts.html` file and used this to extend `footer.html` to point
to (and thus include) the `cookie-consent.js` code.  With this commit,
the cookies banner should appear.

Note that the `loadGAonConsent()` function is missing and needs to be
implemented for the cookie consent for Google Analytics to function as
expected.